pyproject.toml
[project]
name = "hl-bot-v2"
version = "0.1.0"
description = "Hyperliquid Trading Bot V2"
authors = [
    {name = "Guanzhi Ma"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "scipy (>=1.16.0,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "hyperliquid (>=0.4.66,<0.5.0)",
    "hyperliquid-python-sdk (==0.16.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "networkx (>=3.5,<4.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
